Revert "Merge pull request #2280 from kolyshkin/errors-unwrap"
Using errors.Unwrap() is not the best thing to do, since it returns
nil in case of an error which was not wrapped. More to say,
errors package provides more elegant ways to check for underlying
errors, such as errors.As() and errors.Is().

This reverts commit f8e138855d4a630006e88fb7476d35e10e31bf22, reversing
changes made to 6ca9d8e6dae9f5a8f141cf0a3d6a26d74984b2ee.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>